optimized some stuff here and there, refactored code, moved all parallelism stuff to separate files to make it configurable for build.
