+++ .depend/bush.d
@@ -1,6 +1,5 @@
 obj/bush.o: src/bush.c include/bush.h include/tap.h include/fileio.h \
-  include/networks.h include/datastructures.h include/utils.h \
-  include/thpool.h
+  include/networks.h include/datastructures.h include/utils.h
 
 include/bush.h:
 
@@ -13,5 +12,3 @@ include/networks.h:
 include/datastructures.h:
 
 include/utils.h:
-
-include/thpool.h:
