+++ include/bush.h
@@ -40,6 +40,7 @@
 #include "utils.h"
 
 #define NEW_LINK -1
+#define PARALLELISM 1
 
 /*
  * merge_type: Extra data for merge nodes
@@ -229,12 +230,8 @@ void initializeBushesB(network_type *network, bushes_type *bushes,
                        struct algorithmBParameters_type *parameters);
 void updateBushB(int origin, network_type *network, bushes_type *bushes,
                  algorithmBParameters_type *parameters);
-void updateBushB_par(int origin, network_type *network, bushes_type *bushes,
-                     algorithmBParameters_type *parameters);
 void updateFlowsB(int origin, network_type *network, bushes_type *bushes,
                   algorithmBParameters_type *parameters);
-void updateFlowsB_par(int origin, network_type *network, bushes_type *bushes,
-                      algorithmBParameters_type *parameters);
 
 /* Basic bush manipulations */
 bushes_type *createBushes(network_type *network);
@@ -252,39 +249,28 @@ void mergeFirstTopologicalOrder(int origin, network_type *network,
                                 algorithmBParameters_type *parameters);
 void scanBushes(int origin, network_type *network, bushes_type *bushes,
                 algorithmBParameters_type *parameters, bool longestUsed);
-void scanBushes_par(int origin, network_type *network, bushes_type *bushes,
-                    algorithmBParameters_type *parameters, bool longestUsed);
+
 void reconstructMerges(int origin, network_type *network, bushes_type *bushes);
-void reconstructMerges_par(int origin, network_type *network, bushes_type *bushes);
 void findDivergenceNodes(int origin, network_type *network,
                          bushes_type *bushes);
 bool rescanAndCheck(int origin, network_type *network, bushes_type *bushes,
                     algorithmBParameters_type *parameters);
-bool rescanAndCheck_par(int origin, network_type *network, bushes_type *bushes,
-                        algorithmBParameters_type *parameters);
+
 void updateFlowPass(int origin, network_type *network, bushes_type *bushes,
                     algorithmBParameters_type *parameters);
-void updateFlowPass_par(int origin, network_type *network, bushes_type *bushes,
-                        algorithmBParameters_type *parameters);
+
 void calculateBushFlows(int origin, network_type *network,
                         bushes_type *bushes);
-void calculateBushFlows_par(int origin,network_type *network,bushes_type *bushes);
 void pushBackFlowSimple(int j, int origin, network_type *network,
                         bushes_type *bushes);
-void pushBackFlowSimple_par(int j, int origin, network_type *network,
-                            bushes_type *bushes);
+
 void pushBackFlowMerge(merge_type *merge, network_type *network,
                        bushes_type *bushes);
-void pushBackFlowMerge_par(merge_type *merge, network_type *network,
-                           bushes_type *bushes, int t_id);
+
 void rectifyMerge(int j, merge_type *merge, bushes_type *bushes);
-void rectifyMerge_par(int j, merge_type *merge, bushes_type *bushes, int t_id);
 void newtonFlowShift(int j, merge_type *merge, int origin,
                      network_type *network, bushes_type *bushes,
                      algorithmBParameters_type *parameters);
-void newtonFlowShift_par(int j, merge_type *merge, int origin,
-                         network_type *network, bushes_type *bushes,
-                         algorithmBParameters_type *parameters);
 
 /* Utility functions */
 bool isInBush(int origin, int ij, network_type *network, bushes_type *bushes);
@@ -292,11 +278,9 @@ bool isMergeNode(int origin, int i, bushes_type *bushes);
 int pred2merge(int ij);
 int merge2pred(int m);
 void exactCostUpdate(int ij, double shift, network_type *network);
-void exactCostUpdate_par(int ij, double shift, network_type *network);
 void linearCostUpdate(int ij, double shift, network_type *network);
 void noCostUpdate(int ij, double shift, network_type *network);
 void checkFlows(network_type *network, bushes_type *bushes);
-void checkFlows_par(network_type *network, bushes_type *bushes, int t_id);
 
 /**** Merges and merge-doubly linked lists ****/
